Move the google pay environment logic to the params mapper
We need to match the google pay environment and the Adyen environment with each other but only if the merchant doesn't explicitly set the value of the google pay environment. We needed to use a boolean flag for that since everything had to be do in the GooglePayConfiguration.Builder. But now we can move that logic to the params mapper.

COAND-549