Move default configuration values to ComponentParamsMappers
Moving all the default values from the configuration builders to the mappers means that the mappers become the source of truth for all config values, since we already compute other values in there. We need to make sure that we keep the config builder docs up to date though.

COAND-549