Add BaseConfigurationBuilder to itself as a generic type
Base methods (setters for common fields) in BaseConfigurationBuilder need to be able to return the super type of the builder and not BaseConfigurationBuilder, because if we call these methods on a specific builder and they return the base builder, we won't be able to chain any other methods from that specific builder.

Example:

CardConfiguration.Builder()
    .setEnvironment(TEST) // this method used to return BaseConfigurationBuilder
    .setSupportedCardTypes(types) // this method cannot be called because it's defined inside CardConfiguration.Builder not BaseConfigurationBuilder
    .build()

We were solving these by re-overriding all of these common methods in all of the builders and returning the super type, but that generates lots of verbose code and we can easily forget to override these methods.

By adding BuilderT: BaseConfigurationBuilder as a generic type to itself, the builder can have methods that return the generic type which is the type of the super class we need to return.

COAND-549