Merge view flows into a StateFlow
The normal merge function uses a normal flow, which will not keep
state in case of a configuration change. We need to keep the state
to make sure the flow can be continued correctly at all times.

To easily merge flows into a StateFlow the mergeStateFlows was added.
Especially for our view flow use case mergeViewFlows was added.

COAND-660