Refactor PreselectedStoredPaymentMethodFragment
Logic should be moved to the view model, and it should not rely on requiresInput() anymore, but instead we can make use of the component state. When the input is not valid we should show the stored component in a separate screen to allow the shopper to fill the required details (for example CVC). But when the input is valid (for example with setHideCvcStoredCard(true)) the shopper is not required to enter anything so we can skip the next screen step and automatically submit the component from this fragment

COAND-698