Refactor DropInParams to support nullable shopper locale and reduce integration errors
There are multiple sources of truth in drop-in for the amount, configurations, session details etc and the logic is spread across the bundle handler, view model, checkout/drop-in configurations and drop in params. This commit tried to map all the params as early as the DropInViewModelFactory and rely on DropInViewModel.dropInParams as the source of truth for all other classes. It also aligns with the components where we try to remove the configuration from play as early as possible (providers) and rely on one params class for the whole module

COAND-848