Move loading state to GooglePayView
Previously, the GooglePayView was replaced with ProcessingPaymentView. This causes the Activity Result API to fail and not deliver the result in case the view got recreated on configuration change. By keeping the GooglePayView in view it will be able to successfully restore it's state and deliver the activity result. Making the loading state part of the GooglePayView is essential to fix this problem.

COAND-941